#22 Siirretty tasojen jako omaksi tiedostokseen
