+++ src/components/game/GameEngine.js
@@ -4,6 +4,7 @@ import shuffleArray from 'lodash.shuffle';
 import { getPathByName, getWordsForPath } from '../../db/db';
 import { useNavigate } from 'react-router-dom';
 import './GameEngine.css';
+import wordsIntoLevels from './WordsIntoLevels';
 
 const GameEngine = ({ pathName, levelIndex: initialLevelIndex }) => {
   const [currentPhase, setCurrentPhase] = useState(1);
@@ -19,35 +20,6 @@ const GameEngine = ({ pathName, levelIndex: initialLevelIndex }) => {
   const [levelIndex, setLevelIndex] = useState(initialLevelIndex - 1);
 
   const navigate = useNavigate();
-
-  const wordsIntoLevels = (words) => {
-    const totalWords = words.length;
-    let levelCount = 0;
-    // Determine the number of levels (min: 1, max: -)
-    if (totalWords <= 5) {
-      levelCount = totalWords;
-    } else {
-      levelCount = Math.ceil(totalWords / 50) * 5;
-    }
-    // Base number of words per level
-    const baseWordsPerLevel = Math.floor(totalWords / levelCount);
-    // Extra words that need to be added
-    const extraWords = totalWords % levelCount;
-    // Initialize levels
-    const levels = Array.from({ length: levelCount }, () => []);
-    let wordIndex = 0;
-    // Distribute the base number of words to each level
-    for (let i = 0; i < levelCount; i++) {
-      levels[i] = words.slice(wordIndex, wordIndex + baseWordsPerLevel);
-      wordIndex += baseWordsPerLevel;
-    }
-    // Distribute the extra words starting from the first level towards the end
-    for (let i = 0; i < extraWords; i++) {
-      levels[i].push(words[wordIndex]);
-      wordIndex++;
-    }
-    return levels;
-  };
 
   useEffect(() => {
     const fetchData = async () => {
