+++ src/pages/LevelsPage.js
@@ -1,17 +1,24 @@
-// LevelsPage.js
 import React, { useState, useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { getPathByName, getWordsForPath } from '../db/db';
 import PawButton from '../components/levels/PawButton';
 import RunButton from '../components/levels/RunButton';
 import TreeButton from '../components/levels/TreeButton';
+import HouseButton from '../components/levels/HouseButton';
 import BriefCaseButton from '../components/levels/BriefCaseButton';
 import TrophyButton from '../components/levels/TrophyButton';
 import BackButton from '../components/universal/BackButton';
 import Path from '../components/levels/Path';
+import wordsIntoLevels from '../components/game/WordsIntoLevels';
 import './LevelsPage.css';
 
-const levelButtons = [PawButton, RunButton, TreeButton, BriefCaseButton];
+const levelButtons = [
+  PawButton,
+  RunButton,
+  TreeButton,
+  BriefCaseButton,
+  HouseButton,
+];
 const finalButton = TrophyButton;
 const positions = [
   [130, 180],
@@ -27,43 +34,6 @@ const LevelsPage = () => {
   const [levels, setLevels] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
-
-  const wordsIntoLevels = (words) => {
-    const totalWords = words.length;
-    let levelCount = 0;
-
-    // Determine the number of levels (min: 1, max: -)
-    if (totalWords <= 5) {
-      levelCount = totalWords;
-    } else {
-      levelCount = Math.ceil(totalWords / 50) * 5;
-    }
-
-    // Base number of words per level
-    const baseWordsPerLevel = Math.floor(totalWords / levelCount);
-
-    // Extra words that need to be added
-    const extraWords = totalWords % levelCount;
-
-    // Initialize levels
-    const levels = Array.from({ length: levelCount }, () => []);
-
-    let wordIndex = 0;
-
-    // Distribute the base number of words to each level
-    for (let i = 0; i < levelCount; i++) {
-      levels[i] = words.slice(wordIndex, wordIndex + baseWordsPerLevel);
-      wordIndex += baseWordsPerLevel;
-    }
-
-    // Distribute the extra words starting from the first level towards the end
-    for (let i = 0; i < extraWords; i++) {
-      levels[i].push(words[wordIndex]);
-      wordIndex++;
-    }
-
-    return levels;
-  };
 
   useEffect(() => {
     const fetchLevels = async () => {
